+++ platforms/winit/examples/simple.rs
@@ -1,6 +1,4 @@
-use accesskit::{
-    Action, ActionRequest, Live, Node, NodeBuilder, NodeId, Rect, Role, Tree, TreeUpdate,
-};
+use accesskit::{Action, ActionRequest, Live, Node, NodeId, Rect, Role, Tree, TreeUpdate};
 use accesskit_winit::{Adapter, Event as AccessKitEvent, WindowEvent as AccessKitWindowEvent};
 use std::error::Error;
 use winit::{
@@ -33,26 +31,26 @@ const BUTTON_2_RECT: Rect = Rect {
     y1: 100.0,
 };
 
-fn build_button(id: NodeId, name: &str) -> Node {
+fn build_button(id: NodeId, label: &str) -> Node {
     let rect = match id {
         BUTTON_1_ID => BUTTON_1_RECT,
         BUTTON_2_ID => BUTTON_2_RECT,
         _ => unreachable!(),
     };
 
-    let mut builder = NodeBuilder::new(Role::Button);
-    builder.set_bounds(rect);
-    builder.set_name(name);
-    builder.add_action(Action::Focus);
-    builder.add_action(Action::Click);
-    builder.build()
+    let mut node = Node::new(Role::Button);
+    node.set_bounds(rect);
+    node.set_label(label);
+    node.add_action(Action::Focus);
+    node.add_action(Action::Click);
+    node
 }
 
 fn build_announcement(text: &str) -> Node {
-    let mut builder = NodeBuilder::new(Role::Label);
-    builder.set_name(text);
-    builder.set_live(Live::Polite);
-    builder.build()
+    let mut node = Node::new(Role::Label);
+    node.set_value(text);
+    node.set_live(Live::Polite);
+    node
 }
 
 struct UiState {
@@ -69,13 +67,13 @@ impl UiState {
     }
 
     fn build_root(&mut self) -> Node {
-        let mut builder = NodeBuilder::new(Role::Window);
-        builder.set_children(vec![BUTTON_1_ID, BUTTON_2_ID]);
+        let mut node = Node::new(Role::Window);
+        node.set_children(vec![BUTTON_1_ID, BUTTON_2_ID]);
         if self.announcement.is_some() {
-            builder.push_child(ANNOUNCEMENT_ID);
+            node.push_child(ANNOUNCEMENT_ID);
         }
-        builder.set_name(WINDOW_TITLE);
-        builder.build()
+        node.set_label(WINDOW_TITLE);
+        node
     }
 
     fn build_initial_tree(&mut self) -> TreeUpdate {